+++ packages/cmd/src/CommandProcessor.ts
@@ -31,7 +31,7 @@ import { IInvokeCommandParms } from "./doc/parms/IInvokeCommandParms";
 import { ICommandProcessorParms } from "./doc/processor/ICommandProcessorParms";
 import { ImperativeExpect } from "../../expect";
 import { inspect } from "util";
-import { EnvFileUtils, ImperativeConfig, TextUtils } from "../../utilities";
+import { EnvFileUtils, ImperativeConfig, NextVerFeatures, TextUtils } from "../../utilities";
 import * as nodePath from "path";
 import * as os from "os";
 import * as stream from "stream";
@@ -1228,14 +1228,57 @@ export class CommandProcessor {
         if (handlerErr instanceof ImperativeError) {
             this.log.error(`Handler for ${this.mDefinition.name} rejected by thrown ImperativeError.`);
             response.setError(handlerErr.details);
-            response.console.errorHeader("Command Error");
-            response.console.error(Buffer.from(handlerErr.message + "\n"));
-            if ((handlerErr as ImperativeError).details.additionalDetails != null
-                && typeof (handlerErr as ImperativeError).details.additionalDetails === "string") {
-                response.console.errorHeader("Error Details");
-                response.console.error((handlerErr as ImperativeError).details.additionalDetails);
+
+            // display primary user message
+            // TODO:V3_ERR_FORMAT - Don't test for env variable in V3
+            if (NextVerFeatures.useV3ErrFormat()) {
+                response.console.error(TextUtils.chalk.red(
+                    "Unable to perform this operation due to the following problem."
+                ));
+                // Remove http status in 'message', since the same information was placed in additionalDetails.
+                response.console.error(TextUtils.chalk.red(
+                    handlerErr.message.replace(/Rest API failure with HTTP\(S\) status \d\d\d\n/, "")
+                ));
+            } else { // TODO:V3_ERR_FORMAT - Remove in V3
+                response.console.errorHeader("Command Error");
+                response.console.error(Buffer.from(handlerErr.message + "\n"));
             }
+
+            // display server response
+            // TODO:V3_ERR_FORMAT - Don't test for env variable in V3
+            if (NextVerFeatures.useV3ErrFormat()) {
+                const responseTitle = "Response From Service";
+                if (handlerErr.causeErrors) {
+                    try {
+                        const causeErrorsJson = JSON.parse(handlerErr.causeErrors);
+                        response.console.error("\n" + TextUtils.chalk.bold.yellow(responseTitle));
+                        response.console.error(TextUtils.prettyJson(causeErrorsJson, undefined, false, ""));
+                    } catch (parseErr) {
+                        // causeErrors was not JSON.
+                        const causeErrString: string = handlerErr.causeErrors.toString();
+                        if (causeErrString.length > 0) {
+                            // output the text value of causeErrors
+                            response.console.error("\n" + TextUtils.chalk.bold.yellow(responseTitle));
+                            response.console.error(causeErrString);
+                        }
+                    }
+                }
+            }
+
+            // display diagnostic information
+            const diagInfo: string = (handlerErr as ImperativeError).details.additionalDetails;
+            if (diagInfo?.length > 0) {
+                // TODO:V3_ERR_FORMAT - Don't test for env variable in V3
+                if (NextVerFeatures.useV3ErrFormat()) {
+                    response.console.error(TextUtils.chalk.bold.yellow("\nDiagnostic Information"));
+                } else { // TODO:V3_ERR_FORMAT - Remove in V3
+                    response.console.errorHeader("Error Details");
+                }
+                response.console.error(diagInfo);
+            }
+
             response.data.setMessage(handlerErr.message);
+
         } else if (handlerErr instanceof Error) {
             this.log.error(`Handler for ${this.mDefinition.name} rejected by unhandled exception.`);
             response.setError({ msg: handlerErr.message, stack: handlerErr.stack });